Add validators

+++ CalorieTracker.Api/Dtos/FoodEntries/UpdateFoodEntryRequest.cs
@@ -3,6 +3,6 @@
 public record UpdateFoodEntryRequest
 (
     string Barcode,
-    int Grams,
+    decimal Grams,
     DateTime EatenAt
 );
